refactor(backend): :recycle: break create category into smaller files
