+++ backend/backend/main.py
@@ -3,18 +3,20 @@ from fastapi import FastAPI
 from fastapi.middleware.cors import CORSMiddleware
 
 from backend.config import APP_NAME, API_HOST, API_PORT, LOG_LEVEL
-from backend.routers import example, info, authentication, review, create_course, view_my_learning, view_video, study_latest_video
+from backend.routers import course_create, example, info, authentication, review, view_my_learning, view_video, study_latest_video, course, category
 
 app = FastAPI()
 
 app.include_router(example.router)
 app.include_router(info.router)
 app.include_router(authentication.router)
-app.include_router(create_course.router)
+app.include_router(course_create.router)
 app.include_router(view_my_learning.router)
 app.include_router(view_video.router)
 app.include_router(study_latest_video.router)
 app.include_router(review.router)
+app.include_router(course.router)
+app.include_router(category.router)
 
 origins = [
     "http://localhost:3000",
