+++ backend/backend/routers/category.py
@@ -1,0 +1,17 @@
+import uuid
+from fastapi import APIRouter
+
+from backend.controller_instance import controller
+
+
+router = APIRouter()
+
+
+@router.get("/category")
+def get_all_categories():
+    return controller.get_all_categories()
+
+@router.get("/category/{category_id}")
+def get_category_by_id(category_id: str):
+    category = controller.search_category_by_id(uuid.UUID(category_id))
+    return category
